Trimmed unnecessary code from MS/TP receive FSM.  Updated rs485 for AVR and ARM.

+++ bacnet-stack/ports/atmega168/rs485.h
@@ -46,10 +46,9 @@ extern "C" {
 
     void RS485_Transmitter_Enable(bool enable);
 
-    /* NOTE: Only able to send up to 255 bytes at a time */
     void RS485_Send_Data(
         uint8_t * buffer,       /* data to send */
-        uint8_t nbytes);       /* number of bytes of data */
+        uint16_t nbytes);       /* number of bytes of data */
 
     bool RS485_ReceiveError(void);
     bool RS485_DataAvailable(uint8_t *data);
